// src/enums/response.ts
export enum GuardianApiKeys {
    PAGES = "pages",
    URL = "webUrl",
    TITLE = "webTitle",
    PUBLICATION_DATE = "webPublicationDate",
}

export enum NYTimesApiKeys {
    PAGES = 'hits',
    URL = "web_url",
    HEADLINE = "headline",
    TITLE = "main",
    DESCRIPTION = "abstract",
    PUBLICATION_DATE = "pub_date",
    BY_LINE = "byline",
    AUTHER_NAME = "original"
}

export enum NewsApiKey {
    PAGES = "totalResults",
    AUTHER = "author",
    DESCRIPTION = "description",
    PUBLICATION_DATE = "publishedAt",
    TITLE = "title",
    URL = "url",
    IMAGE_URL = "urlToImage"
}

export enum FormattedResultKeys {
    AUTHER = "auther",
    TITLE = "title",
    DESCRIPTION = "description",
    PUBLISHED_AT = "published-at",
    NEWS_URL = "url",
    IMAGE = "image"
}